+++ reload_plot_model.py
@@ -26,6 +26,8 @@ from nilearn.plotting import plot_stat_map
 from nilearn.regions import signals_to_img_labels
 import sys
 
+from eval_utils import test_kl_r2
+
 modelfile = sys.argv[1]
 
 modeldict = torch.load(modelfile)
@@ -35,32 +37,6 @@ modeldict = torch.load(modelfile)
 
 testloader = DataLoader(testset,batch_size=512)
 
-def test_kl_r2(testloader,net,kl_im,kl_audio,kl_places,mseloss):
-    all_fmri = []
-    all_fmri_p = []
-    net.eval()
-    with torch.no_grad():
-        for onesample in testloader:
-
-            bsize = onesample['waveform'].shape[0]
-            
-            # load data
-            wav = torch.Tensor(onesample['waveform']).cuda()
-
-            wav = wav.view(bsize,1,-1,1)
-
-            fmri = onesample['fmri'].cuda()
-
-            # Forward pass
-            _,_,_,fmri_p = net(wav)
-
-            all_fmri.append(fmri.cpu().numpy())
-            all_fmri_p.append(fmri_p.cpu().numpy())
-            
-
-
-    r2_model = r2_score(np.vstack(all_fmri),np.vstack(all_fmri_p),multioutput='raw_values')
-    return r2_model
 
 ### Model Setup
 #net = WaveformCNN(nfeat=nfeat,ninputfilters=ninputfilters,do_encoding_fmri=True)
@@ -84,7 +60,7 @@ print("mean R2 score on test set  : {}".format(r2model.mean()))
 print("max R2 score on test set  : {}".format(r2model.max()))
 
 ### Plot the loss figure
-f = plt.figure()
+f = plt.figure(figsize=(20,10))
 
 ax = plt.subplot(2,1,2)
 
@@ -99,5 +75,7 @@ r2_img = signals_to_img_labels(r2model.reshape(1,-1),'/home/nfarrugi/git/MIST_pa
 ax = plt.subplot(2,1,1)
 
 plot_stat_map(r2_img,display_mode='z',cut_coords=8,figure=f,axes=ax)
-plt.show()
+
 r2_img.to_filename('r2_img.nii.gz')
+f.savefig('test.png')
+plt.close()
